Update haptics.rs
This part is very un-elegant. I actually used ChatGPT to re-write this section, maintaining all the variables, structs, and functions, but reporting haptics working despite not being connected, as well as doing nothing with any haptics command and just reporting 'OK back to whatever calls it. the iWit has no haptics and was failing when it would try and connect. This bypasses the issue but would kill haptics for the Surface. It also changed the vendor and product id it was looking for to that of the iWit. Changing the Product and vendor id variables to a vector and iterating over them would allow both to be detected, and one could more than likely make case statements for haptics-supported vs. not.